Counter nlp update (#3588)
* add eval for spark computtationgraph

* added new transformprocess call

* make parallelwrappermain emeddable within another main

* local

* remove old parallelwrapper invocations /cc @raver119

* update nn server to have scala version

* add knn for data point

* get rid of org.json dep

* change vptree storage to float

* updated float functions using counter in nlp modules

+++ deeplearning4j-nlp-parent/deeplearning4j-nlp/src/main/java/org/deeplearning4j/models/embeddings/learning/impl/elements/GloVe.java
@@ -283,7 +283,7 @@ public class GloVe<T extends SequenceElement> implements ElementsLearningAlgorit
                         continue;
                     }
 
-                    errorCounter.incrementCount(epochId, iterateSample(element1, element2, weight));
+                    errorCounter.incrementCount(epochId, (float) iterateSample(element1, element2, weight));
                     if (pairsCounter.incrementAndGet() % 1000000 == 0) {
                         log.info("Processed [" + pairsCounter.get() + "] word pairs so far...");
                     }
